sick jwt integration

+++ src/compiler.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 cd $( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+. bash_colors.sh
 VENV=".venv"
 BORG_NAME="MYBORG"
 PLAINTEXT_PASSPHRASE="12345678"
@@ -26,29 +27,69 @@ doSetup(){
     pip install -r ../requirements.d/development.txt >/dev/null
     pip install -e ../ >/dev/null
 }
-doTestBorg(){
 
+doTestBorg(){
+    if [[ "$1" == "" ]]; then
+        echo "[doTestBorg] : missing argument"
+        exit 1
+    fi
+    clr_green Testing borg with $1
+    env | egrep "BORG_|AUTH_TOKEN|PUBLIC_KEY|AUDIENCE"
     $1 --version
 
     echo 1234 > testfile.txt
     rm -rf test.borg
-    BORG_PASSPHRASE="$PLAINTEXT_PASSPHRASE" $1 init -e repokey test.borg
+    $1 init -e repokey test.borg
     ls -al test.borg
 
-    echo; echo Borging with plaintext passphrase; echo
-    BORG_PASSPHRASE="$PLAINTEXT_PASSPHRASE" $1 create test.borg::test1 testfile.txt
-    BORG_PASSPHRASE="$PLAINTEXT_PASSPHRASE" $1 list test.borg::test1
-    echo; echo OK; echo
-
-    echo; echo Borging with encrypted passphrase; echo
-    BORG_PASSPHRASE_ENCRYPTED="$ENCRYPTED_PASSPHRASE" $1 create test.borg::test2 testfile.txt
-    BORG_PASSPHRASE_ENCRYPTED="$ENCRYPTED_PASSPHRASE" $1 list test.borg::test2
-    echo; echo OK; echo
+    echo; clr_green Borging with plaintext passphrase; echo
+    $1 create test.borg::test1 testfile.txt
+    $1 list test.borg::test1
+    echo; clr_green OK; echo
 
     rm -rf test.borg
 }
 
+doPyInstaller(){
 
+
+    pyinstaller \
+        --onedir \
+        -y \
+        --clean \
+        -n $BORG_NAME \
+            borg/__main__.py
+
+}
+
+
+
+
+
+doPassphraseTests(){
+
+    clr_green Testing Borg with $BP
+    file $BP
+
+    BORG_PASSPHRASE="$PLAINTEXT_PASSPHRASE" \
+        doTestBorg $BP
+    echo; clr_green Plaintext with specified passphrase tests pass; echo; echo
+    unset BORG_PASSPHRASE
+
+
+    AUDIENCE="$(cat ~/.keys/audience.key)" \
+    PUBLIC_KEY="$(cat ~/.keys/pub.key|base64 -w0)" \
+    AUTH_TOKEN="$(generateApplicationToken.sh apiStatus read)" \
+        doTestBorg $BP
+    echo; clr_green Plaintext passphrase in AUTH_TOKEN tests pass; echo
+
+    AUDIENCE="$(cat ~/.keys/audience.key)" \
+    PUBLIC_KEY="$(cat ~/.keys/pub.key|base64 -w0)" \
+    AUTH_TOKEN="$(generateApplicationToken.sh araAPI_ro read)" \
+        doTestBorg $BP
+    echo; clr_green Encrypted passphrase in AUTH_TOKEN tests pass; echo
+
+}
 
 doSetup
 
@@ -56,30 +97,20 @@ doSetup
 
 
 BP="$VENV/bin/borg"
-doTestBorg $BP
-echo; echo Plaintext tests pass; echo
+doPassphraseTests
 
 
-exit
 
-
-pyinstaller \
-    --onedir \
-    -y \
-    --clean \
-    -n $BORG_NAME \
-        borg/__main__.py
-
-echo; echo OK; echo
+doPyInstaller
 
 
 
 BP="./dist/$BORG_NAME/$BORG_NAME"
+doPassphraseTests
 
-doTestBorg $BP
+echo; clr_green Pyinstaller tests pass; echo
 
-echo; echo Pyinstaller tests pass; echo
-
+exit
 
 #pyarmor register \
 #    ~/.pyarmor/pyarmor-regfile-1.zip
